+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unrealeased]
 
+### Added
+
+- Add catalog visibility field to course run so it's possible to hide a
+  course run on the catalog and/or hide it completely. Possible values
+  are `course_and_search`, `course_only` and `hidden`.
+
 ### Changed
 
 - Do not use absolute positioning to place course glimpse metadata